+++ apps/web/src/app.test.tsx
@@ -197,6 +197,7 @@ const clientWith = (
   updateSession: () => Promise.reject(new Error('updateSession is not stubbed in this file')),
   withdrawSession: () => Promise.reject(new Error('withdrawSession is not stubbed in this file')),
   restoreSession: () => Promise.reject(new Error('restoreSession is not stubbed in this file')),
+  mergeSession: () => Promise.reject(new Error('mergeSession is not stubbed in this file')),
   getMeals: () => Promise.resolve({ intro_markdown: '', slots: [], meals: [] }),
   updateMeal: () => Promise.reject(new Error('updateMeal is not stubbed in this file')),
   setMealLead: () => Promise.reject(new Error('setMealLead is not stubbed in this file')),